feat: Support building the CLI for js/wasm
The CLI did not compile for GOOS=js GOARCH=wasm, so the dashboard
could not run it in the browser. The build graph reached BuildKit,
containerd, Docker, Bubble Tea, a machine fingerprint and a large
generated IATA table, none of which the wasm backend accepts.

Commands that need a local daemon, terminal or filesystem now split
into `!js` and `js` files. The js side keeps the kong structs, so
help output is unchanged, and reports that the command belongs to the
native CLI. A `wasm` task builds and compresses the module.
