+++ internal/cmd/build_js.go
@@ -1,0 +1,25 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2026, Unikraft GmbH and The Unikraft CLI Authors.
+// Licensed under the BSD-3-Clause License (the "License").
+// You may not use this file except in compliance with the License.
+
+//go:build js
+
+package cmd
+
+import (
+	"context"
+
+	"unikraft.com/cli/internal/config"
+	"unikraft.com/cli/internal/resource"
+)
+
+// Run reports that image builds need a local container builder.
+func (c *ImageBuildCmd) Run(ctx context.Context, cfg *config.Config, partition *resource.Partition) error {
+	return notAvailable("image build")
+}
+
+// Run reports that image builds need a local container builder.
+func (c *BuildCmd) Run(ctx context.Context, cfg *config.Config, partition *resource.Partition) error {
+	return notAvailable("build")
+}
