+++ internal/cmd/instances.go
@@ -19,7 +19,6 @@ import (
 	"slices"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/MakeNowJust/heredoc"
 	"github.com/distribution/reference"
@@ -42,7 +41,6 @@ import (
 	"unikraft.com/cli/internal/resource/cmd"
 	"unikraft.com/cli/internal/resource/value"
 	"unikraft.com/cli/internal/timeouts"
-	"unikraft.com/cli/internal/tunnel"
 	"unikraft.com/cli/internal/types"
 )
 
@@ -2265,34 +2263,4 @@ func (cmd InstancesTunnelCmd) Examples() []kingkong.Example {
 			Commands:    []string{"unikraft instance tunnel -p 5500 my-instance:8080"},
 		},
 	}
-}
-
-func (cmd *InstancesTunnelCmd) Run(ctx context.Context, stdio config.Stdio) error {
-	targets, err := tunnel.ParseTargets(cmd.Targets, cmd.TunnelProxyPorts)
-	if err != nil {
-		return fmt.Errorf("could not parse targets: %w", err)
-	}
-
-	tun, err := tunnel.New(ctx, targets)
-	if err != nil {
-		return fmt.Errorf("could not create tunnel: %w", err)
-	}
-
-	g, err := multimetro.NewClient(ctx)
-	if err != nil {
-		return err
-	}
-
-	defer func() {
-		// Detach from ctx so cleanup survives the same Ctrl-C that triggered
-		// it, but bound it so a stuck API call can't leave a still-billed
-		// proxy instance orphaned.
-		closeCtx, cancel := context.WithTimeout(context.WithoutCancel(ctx), 30*time.Second)
-		defer cancel()
-		if err := tun.Close(closeCtx, g); err != nil {
-			log.G(ctx).Error().Err(err).Msg("could not terminate tunnel proxy")
-		}
-	}()
-
-	return tun.Run(ctx, g, cmd.ProxyControlPort, cmd.TunnelServiceImage)
 }
