+++ internal/images/images.go
@@ -3,35 +3,25 @@
 // Licensed under the BSD-3-Clause License (the "License").
 // You may not use this file except in compliance with the License.
 
+//go:build !js
+
 package images
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/containerd/containerd/v2/core/remotes/docker"
-	"github.com/distribution/reference"
 	imagespec "unikraft.com/x/image-spec"
 
 	"unikraft.com/cli/internal/config"
-	xreference "unikraft.com/cli/internal/x/reference"
 )
-
-const DefaultRegistry = "unikraft.io"
 
 var defaultRegistries = []string{
 	"unikraft.io",
 	"index.unikraft.io",
 }
 
-type insecureContextKey struct{}
-
-// WithInsecureContext returns a context that carries insecure registry options,
-// which are picked up by Accessor.
-func WithInsecureContext(ctx context.Context, opts ...AccessorOpt) context.Context {
-	return context.WithValue(ctx, insecureContextKey{}, opts)
-}
-
+// Accessor builds a registry accessor for the current profile.
 func Accessor(ctx context.Context, opts ...AccessorOpt) (*imagespec.Accessor, error) {
 	if len(opts) == 0 {
 		if ctxOpts, ok := ctx.Value(insecureContextKey{}).([]AccessorOpt); ok {
@@ -58,55 +48,4 @@ func Accessor(ctx context.Context, opts ...AccessorOpt) (*imagespec.Accessor, er
 		imagespec.WithRegistryHeaders(options.Headers),
 		imagespec.WithReferenceParser(ParseNormalizedNamed),
 	), nil
-}
-
-// AccessorOpt is a functional option for configuring an Accessor.
-type AccessorOpt func(*accessorOpts)
-
-type accessorOpts struct {
-	insecureRegistries []string
-	allInsecure        bool
-}
-
-func WithInsecureRegistry(hosts ...string) AccessorOpt {
-	return func(o *accessorOpts) {
-		o.insecureRegistries = hosts
-	}
-}
-
-func WithInsecureRegistries() AccessorOpt {
-	return func(o *accessorOpts) {
-		o.allInsecure = true
-	}
-}
-
-func ParseNormalizedNamed(key string) (reference.Named, error) {
-	return ParseNormalizedNamedMetro(nil, key)
-}
-
-func ParseNormalizedNamedMetro(metro *config.Metro, key string) (reference.Named, error) {
-	if uri, err := imagespec.ParseURI(key); err == nil {
-		if uri.Scheme != imagespec.URISchemeOCI {
-			return nil, fmt.Errorf("%w: invalid scheme %q", reference.ErrReferenceInvalidFormat, uri.Scheme)
-		}
-		key = uri.Path
-	}
-
-	index := DefaultRegistry
-	if metro != nil {
-		index = metro.Index().Host
-	}
-	return xreference.ParseNormalizedNamed(
-		key,
-		xreference.WithDefaultDomain(index),
-		xreference.WithDefaultPrefix("official/"),
-	)
-}
-
-func FamiliarString(ref reference.Reference) string {
-	return xreference.FamiliarString(
-		ref,
-		xreference.WithDefaultDomain(DefaultRegistry),
-		xreference.WithDefaultPrefix("official/"),
-	)
 }
