+++ internal/telemetry/telemetry.go
@@ -20,7 +20,6 @@ import (
 
 	"github.com/posthog/posthog-go"
 
-	"unikraft.com/x/fingerprint"
 	"unikraft.com/x/version"
 )
 
@@ -81,29 +80,6 @@ func Init() error {
 	sessionID = generateSessionID()
 
 	return nil
-}
-
-// generateDistinctID creates an anonymous distinct ID based on machine fingerprint.
-// The ID is a SHA-256 hash to ensure privacy while maintaining consistency.
-func generateDistinctID() string {
-	fp, err := fingerprint.New()
-	if err != nil {
-		// Fallback to hostname-based ID
-		hostname, _ := os.Hostname()
-		hash := sha256.Sum256([]byte(hostname + "-unikraft-cli"))
-		return hex.EncodeToString(hash[:16])
-	}
-
-	// Create a stable fingerprint string from machine characteristics
-	fpStr := fmt.Sprintf("%s-%s-%s-%s-%t",
-		fp.Hostname,
-		fp.Os,
-		fp.Goarch,
-		fp.Goos,
-		fp.Container,
-	)
-	hash := sha256.Sum256([]byte(fpStr))
-	return hex.EncodeToString(hash[:16])
 }
 
 // generateDistinctID creates a unique session ID for this CLI invocation, which
